Finish Change Email API

+++ public/index.php
@@ -1,6 +1,7 @@
 <?php
 
 use InteractivePlus\PDK2021\Controllers\ReturnableResponse;
+use InteractivePlus\PDK2021\Controllers\UserSystem\LoggedInFunctionController;
 use InteractivePlus\PDK2021\Controllers\UserSystem\LoginController;
 use InteractivePlus\PDK2021\Controllers\UserSystem\RegisterController;
 use InteractivePlus\PDK2021\Controllers\VeriCode\VeriCodeController;
@@ -75,16 +76,19 @@ $customErrorHandler = function(
     return $response;
 };
 
+$errMiddleWare->setDefaultErrorHandler($customErrorHandler);
+
 
 $app->post('/user',RegisterController::class . ':register');
 $app->post('/user/token',LoginController::class . ':login');
-$app->get('/user/token/checkTokenResult',LoginController::class . ':checkTokenValid');
-$app->get('/user/token/refreshResult',LoginController::class . ':refreshToken');
-$app->delete('/user/token/{access_token}',LoginController::class . ':logout');
+$app->get('/user/{uid}/token/{access_token}/checkTokenResult',LoginController::class . ':checkTokenValid');
+$app->get('/user/{uid}/token/refreshResult',LoginController::class . ':refreshToken');
+$app->delete('/user/{uid}/token/{access_token}',LoginController::class . ':logout');
 $app->get('/vericodes/verifyEmailResult/{veriCode}',VeriCodeController::class . ':verifyEmail');
 $app->get('/vericodes/verifyPhoneResult/{veriCode}',VeriCodeController::class . ':verifyPhone');
 $app->post('/vericodes/sendAnotherVerifyEmailRequest',VeriCodeController::class . ':requestVerificationEmailResend');
 $app->post('/vericodes/sendAnotherVerifyPhoneRequest',VeriCodeController::class . ':requestVerificationPhoneResend');
-
+$app->post('/vericodes/changeEmailAddrRequest',VeriCodeController::class . ':requestChangeEmailAddressVeriCode');
+$app->patch('/user/email',LoggedInFunctionController::class . ':changeEmailAddress');
 
 $app->run();
